css/css-fonts/variations/font-style-parsing-shorthand.html

<!DOCTYPE html>
<html>
<head>
    <title>Testing new font-style values introduced in CSS Fonts level 4 (in font shorthand)</title>
    <link rel="help" href="https://www.w3.org/TR/css-fonts-4/#font-style-prop" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="shorthand-test">Shorthand test</div>

    <script>

        var testFontShorthand = [
            { value: "italic 500 24pt Arial",        isValid:true,  expectedStyle: "italic",        expectedWeight:"500", message: "'italic' followed by valid weight" },
            { value: "oblique 50 24pt Arial",        isValid:true,  expectedStyle: "oblique",       expectedWeight:"50",  message: "'oblique' followed by valid small weight" },
            { value: "oblique 500 24pt Arial",       isValid:true,  expectedStyle: "oblique",       expectedWeight:"500", message: "'oblique' followed by valid large weight" },
            { value: "oblique 45deg 500 24pt Arial", isValid:true,  expectedStyle: "oblique 45deg", expectedWeight:"500", message: "'oblique' with positive angle followed by valid weight" },
            { value: "oblique -45deg 500 24pt Arial", isValid:true,  expectedStyle: "oblique -45deg", expectedWeight:"500", message: "'oblique' with negative angle followed by valid weight" }
        ];

        testFontShorthand.forEach(function (testCase) {
            test(() => {
                assert_equals(window.CSS.supports("font", testCase.value), testCase.isValid, "Font shorthand: " + testCase.message);

                var testElement = document.getElementById("shorthand-test");
                testElement.setAttribute("style", "font:" + testCase.value);
                var style = window.getComputedStyle(testElement);
                assert_equals(style.fontStyle, testCase.expectedStyle, "Font shorthand expected style: " + testCase.message);
                assert_equals(style.fontWeight, testCase.expectedWeight, "Font shorthand expected weight: " + testCase.message);
            }, "Font shorthand: " + testCase.message);

        });

    </script>
</body>
</html>
